requirements.txt
# Sentinel-Graph Engine Dependencies

# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
pyyaml>=6.0
faker>=18.0.0

# Graph
networkx>=3.1
python-louvain>=0.16  # community detection (import community)

# Machine Learning
scikit-learn>=1.3.0
torch>=2.0.0
torch-geometric>=2.4.0

# Visualization & Dashboard
streamlit>=1.28.0
plotly>=5.15.0
umap-learn>=0.5.4
folium>=0.14.0
streamlit-folium>=0.15.0

# Geospatial
geopy>=2.3.0

# Privacy
diffprivlib>=0.6.0

# API / Web server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
pydantic>=2.0.0
